refactor: log imports e package
