+++ src/main/java/nttdata/bank/service/ports/ClienteExternoService.java
@@ -1,4 +1,4 @@
-package nttdata.bank.service.adapters;
+package nttdata.bank.service.ports;
 
 import org.springframework.stereotype.Service;
