+++ src/main/java/nttdata/bank/service/transacao/TransacaoService.java
@@ -6,8 +6,8 @@ import nttdata.bank.domain.entities.transacao.TipoDespesaEnum;
 import nttdata.bank.domain.entities.transacao.TipoTransacaoFinEnum;
 import nttdata.bank.domain.entities.transacao.Transacao;
 import nttdata.bank.repository.transacao.TransacaoRepository;
-import nttdata.bank.service.adapters.ClienteExternoService;
-import nttdata.bank.service.adapters.ConverteCambioService;
+import nttdata.bank.service.ports.ClienteExternoService;
+import nttdata.bank.service.ports.ConverteCambioService;
 import nttdata.bank.service.conta.ContaService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
